requirements.txt
ace==0.3.3
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
Django==5.0.4
django-bootstrap4==24.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
Faker==25.4.0
filelock==3.14.0
fonttools==4.53.0
idna==3.7
importlib==1.0.4
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
speedtest-cli==2.1.3
sqlparse==0.5.0
stack-data==0.6.3
threadpoolctl==3.5.0
tk==0.1.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.1
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.0
